requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
elasticsearch==8.11.0
neo4j==5.16.0
networkx==3.2.1
python-dotenv==1.0.0
requests
faiss-gpu-cu12
numpy
pandas
tqdm==4.67.1
sentence_transformers
transformers>=4.44.0
pymysql
neo4j